Fix for previous commit and escaped char values > 127 (0x7F)

+++ src/xpdev/genwrap.c
@@ -246,7 +246,7 @@ char* c_escape_str(const char* src, char* dst, size_t maxlen, BOOL ctrl_only)
 			strncpy(d,e,maxlen-(d-dst));
 			d+=strlen(d);
 		} else if((uchar)*s < ' ' || (uchar)*s >= '\x7f') {
-			d += safe_snprintf(d, maxlen-(d-dst), "\\x%02X", *s);
+			d += safe_snprintf(d, maxlen-(d-dst), "\\x%02X", (uchar)*s);
 		} else *d++=*s;
 	}
 	*d=0;
